[BACKLOG-26664] Moving pentaho/type/action to pentaho/action
