+++ assemblies/webpackage/src/main/resources-filtered/common-ui/package.json
@@ -21,6 +21,7 @@
     "pentaho/lang": "/pentaho/lang",
     "pentaho/i18n": "/pentaho/i18n",
     "pentaho/data": "/pentaho/data",
+    "pentaho/action": "/pentaho/action",
     "pentaho/type": "/pentaho/type",
     "pentaho/visual": "/pentaho/visual",
     "pentaho/config": "/pentaho/config",
@@ -168,9 +169,6 @@
       },
       "pentaho/type/mixins/Enum": {
         "alias": "enum",
-        "base": "element"
-      },
-      "pentaho/type/action/Base": {
         "base": "element"
       },
       "pentaho/data/filter/Abstract": {
@@ -327,20 +325,6 @@
       },
       "pentaho/visual/models/Bubble": {
         "base": "pentaho/visual/models/Abstract"
-      },
-      "pentaho/visual/action/Base": {
-        "base": "pentaho/type/action/Base"
-      },
-      "pentaho/visual/action/Select": {
-        "alias": "select",
-        "base": "pentaho/visual/action/Base"
-      },
-      "pentaho/visual/action/Execute": {
-        "alias": "execute",
-        "base": "pentaho/visual/action/Base"
-      },
-      "pentaho/visual/action/Update": {
-        "base": "pentaho/visual/action/Base"
       },
       "pentaho/visual/color/Palette": {
         "base": "complex"
