+++ assemblies/webpackage/src/main/resources-filtered/common-ui/preconfig.js
@@ -105,6 +105,7 @@ if(!useDebug) {
     "pentaho/lang",
     "pentaho/i18n",
     "pentaho/data",
+    "pentaho/action",
     "pentaho/type",
     "pentaho/visual",
     "pentaho/config",
@@ -361,7 +362,7 @@ if(!useDebug) {
     "pentaho/visual/role/Mapping",
     "pentaho/visual/role/Property",
     "pentaho/visual/base/Model",
-    "pentaho/type/action/Base",
+    "pentaho/action/Base",
     "pentaho/visual/action/Base",
     "pentaho/visual/action/Update",
     "pentaho/visual/action/mixins/Data",
@@ -369,10 +370,11 @@ if(!useDebug) {
     "pentaho/visual/action/SelectionModes",
     "pentaho/visual/action/Select",
     "pentaho/visual/action/Execute",
-    "pentaho/type/action/States",
+    "pentaho/action/States",
     "pentaho/lang/RuntimeError",
-    "pentaho/type/action/Execution",
-    "pentaho/type/action/impl/Target",
+    "pentaho/action/Execution",
+    "pentaho/action/impl/Target",
+    "pentaho/visual/base/ModelChangedError",
     "pentaho/util/BitSet",
     "pentaho/visual/base/View",
     "pentaho/visual/role/ExternalMapping",
