+++ impl/client/src/main/doc-js/pentaho/action/impl/_namespace.jsdoc
@@ -1,5 +1,5 @@
 /*!
- * Copyright 2010 - 2017 Hitachi Vantara. All rights reserved.
+ * Copyright 2017 Hitachi Vantara. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,9 +15,9 @@
  */
 
 /**
- * @name pentaho.type.action.spec
+ * @name pentaho.action.impl
  * @namespace
  *
- * @description The `action.spec` namespace contains specification interfaces and type definitions.
+ * @description The `action.impl` namespace contains implementations related with the action concept.
  */
