+++ impl/client/src/main/doc-js/pentaho/action/spec/IBase.jsdoc
@@ -15,23 +15,34 @@
  */
 
 /**
- * The `spec.IBaseType` interface represents the specification of an action type.
+ * The `spec.IBase` interface describes an action instance using a generic, _object_ form.
  *
- * @name pentaho.type.action.spec.IBaseType
+ * @name pentaho.action.spec.IBase
+ *
  * @interface
- * @extends pentaho.type.spec.IValueType
+ * @extends pentaho.type.spec.IValue
  */
 
 /**
- * Indicates if the action is synchronous.
+ * The label of the action instance.
  *
- * When unspecified, inherits the value of the ancestor action type.
- * The root action type's value is `true`.
+ * When unspecified,
+ * the label of the action type,
+ * {@link pentaho.type.Type#label}, is used.
  *
- * If a {@link Nully} value is specified, it is ignored.
- * If a non-boolean value is specified, it is converted to boolean by passing it to the `Boolean` function.
+ * @name label
+ * @memberOf pentaho.action.spec.IBase#
+ * @type {nonEmptyString}
+ */
+
+/**
+ * The description of the action instance.
  *
- * @name isSync
- * @memberOf pentaho.type.action.spec.IBaseType#
- * @type {boolean}
+ * When unspecified,
+ * the description of the action type,
+ * {@link pentaho.type.Type#description}, is used.
+ *
+ * @name description
+ * @memberOf pentaho.action.spec.IBase#
+ * @type {nonEmptyString}
  */
