+++ impl/client/src/main/doc-js/pentaho/visual/action/spec/IBase.jsdoc
@@ -21,5 +21,5 @@
  * @name pentaho.visual.action.spec.IBase
  *
  * @interface
- * @extends pentaho.type.action.spec.IBase
+ * @extends pentaho.action.spec.IBase
  */
